src/requirements.txt
aiohttp==3.7.3
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.12.5
chardet==3.0.4
configobj==5.0.6
Django==3.1.7
djangorestframework==3.12.2
ecdsa==0.16.1
idna==2.10
idna-ssl==1.1.0
multidict==5.1.0
pbkdf2==1.3
pkg-resources==0.0.0
psycopg2==2.8.6
pyaes==1.6.1
pyblake2==1.1.2
pycurl==7.43.0.6
python-dotenv==0.15.0
pytz==2021.1
requests==2.25.1
sentry-sdk==0.20.3
six==1.15.0
slackclient==2.9.3
slick-bitcoinrpc==0.1.4
sqlparse==0.4.1
typing-extensions==3.7.4.3
ujson==4.0.2
urllib3==1.26.3
yarl==1.6.3
